requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
SQLAlchemy
asyncpg
openai
pinecone-client
PyMuPDF
psycopg2-binary
alembic
